requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
Faker==37.1.0
idna==3.10
python-dotenv==1.1.0
requests==2.32.3
robotframework==7.2.2
robotframework-requests==0.9.7
tzdata==2025.2
urllib3==2.3.0